pyproject.toml
[project]
name = "stock-valuation-app"
version = "0.1.0"
description = "A single-page web app displaying stock data, growth charts, and valuation insights."
readme = "README.md"
authors = [
    { name = "Dimeji Salau", email = "dimejisalau@protonmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "dash>=2.18.2",
    "duckdb>=1.1.3",
    "fastapi>=0.115.4",
    "httpx>=0.27.2",
    "polars>=1.12.0",
    "pyarrow>=18.0.0",
    "pydantic>=2.9.2",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
]

[project.scripts]
stock-valuation-app = "stock_valuation_app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "mypy>=1.13.0",
    "pytest>=8.3.3",
    "ruff>=0.7.3",
]
